requirements.txt
dash>=2.15
dash-bootstrap-components>=1.5
plotly>=5.18
pandas>=2.1
requests>=2.31
cachetools>=5.3
gunicorn>=21.2
python-dotenv>=1.0
